Slack_ingestion/slack_pathway/pyproject.toml
[project]
name = "slack-pathway"
version = "0.1.0"
description = "Real-time RAG system for hackathon chat monitoring"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pathway (>=0.26.1,<0.27.0)",
    "google-generativeai (>=0.3.0,<1.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "pandas (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
